Added Vue and CSS Reset
Vue was missing in clean Laravel build. Added it to dependencies in
package.json.
Added CSS reset to reduce browser inconsistencies in
resources/sass/css-reset. CSS reset was added in a separate file to
make SCSS more readable and not clogg main file. CSS reset is imported
in app.scss.

 #2